pyproject.toml
[project]
name = "learning-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.23.1",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.10.0",
    "openai>=1.70.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
